% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_order_rev.R
\name{fact_order_rev}
\alias{fact_order_rev}
\title{Sorting factor levels in reverse order}
\usage{
fact_order_rev(x)
}
\arguments{
\item{Factor}{}
}
\value{
Factor
}
\description{
Sorting factor levels in reverse order
}
\examples{
fact_order_rev(factor(c("short", "long", "short", "medium", "medium")))
}

