% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factRordering-package.r
\docType{package}
\name{factRordering}
\alias{factRordering}
\alias{factRordering-package}
\title{factRordering. The package contains a function which reverses the input parameters.}
\description{
factRordering. The package contains a function which reverses the input parameters.
}

